Workaround for bug in Monolog
Monolog is expecting an indexed array and ObjectManager passes it an associative array. This sets the key names to numbers which are handled in the same fashion as an indexed array. This got patched in Monolog but not sure when Magento (currently targeting Monolog 1.16) will pull in the updated library.