#421_41 - Accommodate shipping discount in tax calculation
\Magento\Tax\Model\Calculation\AbstractAggregateCalculator::calculateWithTaxNotInPrice
makes a call to get the discount amount $item->getDiscountAmount()
but the shipping item returns a null value for the discount amount causing it to calculate tax on the shipping item regardless.

The items that are passed into the calculation routines come from
\Magento\Tax\Model\Sales\Total\Quote\Tax::collect
\Magento\Tax\Model\Sales\Total\Quote\Tax::getQuoteTaxDetails

Starts building $itemDataObjects for normal quote items via $this->mapItems()
\Magento\Tax\Model\Sales\Total\Quote\CommonTaxCollector::mapItems
which is builds from getting the details out of the ShippingAssignmentInterface item data.

It then gets the shipping item via $this->getShippingDataObject()
\Magento\Tax\Model\Sales\Total\Quote\CommonTaxCollector::getShippingDataObject
and it builds the data from the details in QuoteAddress\Total details that are passed in.
It looks for $total->getShippingDiscountAmount() but that value isn't set.

When the quote discounts are collected
\Magento\SalesRule\Model\Quote\Discount::collect
$this->calculator->processShippingAmount($address)
\Magento\SalesRule\Model\Validator::processShippingAmount
There ends up being a shipping discount amount on the address object, but it never gets set to the total object.

The shipping discount amount does end up being accessible from the ShippingAssignmentInterface object
that is passed into the collect() methods during tax calculation though:
$shippingAssignment->getShipping()->getAddress()->getShippingDiscountAmount()