AVS-421_116 - Start structure and functionality for AsscociatedTaxables
Associated Taxables are used by Gift Wrapping, WEEE Tax, and some 3rd
party extension to allow calculcating tax for orders on items that are
not technically "items" in the Magento sense. This commit begins the
basic functionality of persisting the Associated Taxable information
from the order step so it can be re-transmitted to Avalara at the time
of invoice processing.