Update to TaxClass.php
When using products that make use of the "Customizable Options" with values in "Sku" field. when calling the product->getSku() you get a sku with all selected options' sku tacked on with the product sku. then in productRepository->get($sku) the search for the sku fails and the user cannot add the product to the cart.