Update getting-started.md
Adding in the remaining tables that have to be deleted as well as entries in the core_config_data table that need to be removed in order to delete all AvaTax settings. If the core_config_data items are not removed, then upon re-installation of the extension it will remember all settings. In the case of needing to reset the Taxation Policy, this is the only way to do it.